backend/requirements.txt
Django>=4.0.0
djangorestframework>=3.13.0
django-cors-headers>=3.10.0
Pillow>=9.0.0
gradio-client>=1.10.0
huggingface-hub>=0.19.3